PostCard component finalized
